[iOS] MOVE: DataModel 이름 정리 #189
