+++ src/iOS/KickTube/KickTube/Features/Home/DataModel/DomainModel/CreateRoomDomainModel.swift
@@ -7,6 +7,8 @@
 
 import Foundation
 
+import ManipulateDataModel
+
 struct CreateRoomDomainModel: DTOMappable {
     let code: String
