+++ src/iOS/KickTube/KickTube/Features/Login/DataModel/Request/LoginRequestDTO.swift
@@ -1,0 +1,12 @@
+//
+//  LoginRequestDTO.swift
+//  KickTube
+//
+//  Created by 김수경 on 2/17/25.
+//
+
+import Foundation
+
+struct LoginRequestDTO: Encodable {
+    let device = "mobile"
+}
