+++ src/iOS/KickTube/KickTube/Features/Login/DataModel/Response/TokenResponseDTO.swift
@@ -1,0 +1,17 @@
+//
+//  TokenResponseDTO.swift
+//  KickTube
+//
+//  Created by 김수경 on 2/17/25.
+//
+
+import Foundation
+
+import ManipulateDataModel
+
+@DecodeDTO
+@ConvertToDomainModel<TokenDomainModel>
+struct TokenResponseModel {
+    let accessToken: String
+    let refreshToken: String
+}
