--- src/iOS/KickTube/KickTube/Features/Profile/DataModel/UserProfileResponse.swift
@@ -1,53 +1,0 @@
-//
-//  UserProfileResponse.swift
-//  KickTube
-//
-//  Created by 김수경 on 2/10/25.
-//
-
-import Foundation
-
-import ManipulateDataModel
-
-@DecodeDTO
-@ConvertToDomainModel<UserProfileDomainModel>
-struct UserProfileResponse {
-    @Key("userId") let userID: Int
-    let email: String
-    let nickname: String
-    let role: Int
-    @Key("profileImageUrl") let profileImageURL: String?
-    let profileImages: [String]?
-    let stateMessage: String?
-}
-
-struct UserProfileDomainModel: DTOMappable {
-    let userID: Int
-    let email: String
-    let nickname: String
-    let role: Int
-    let profileImageURL: String?
-    let profileImages: [String]?
-    let stateMessage: String?
-    
-    func toModel() -> UserProfileViewModel {
-        .init(userID: self.userID,
-              email: self.email,
-              nickname: self.nickname,
-              profileImageURL: self.profileImageURL,
-              stateMessage: self.stateMessage)
-    }
-}
-
-
-import UIKit
-
-struct UserProfileViewModel: Codable {
-    let userID: Int
-    let email: String
-    let nickname: String
-    let profileImageURL: String?
-    let stateMessage: String?
-    
-    var profileImageData: Data? = UIImage.profile.toData()
-}
